Rename FreeText block to TextQuestion
