+++ packages/block-editor/src/index.js
@@ -1,6 +1,6 @@
 export { default as answerBlock } from './answer';
-export { default as freeTextQuestionBlock } from './free-text';
 export { default as multipleChoiceQuestionBlock } from './multiple-choice';
 export { default as pollBlock } from './poll';
 export { default as pollAnswerBlock } from './poll-answer';
 export { default as submitButtonBlock } from './submit-button';
+export { default as textQuestionBlock } from './text-question';
