+++ packages/blocks/src/index.js
@@ -1,10 +1,10 @@
 export * from './components';
 
 export { default as Answer } from './answer';
-export { default as FreeText } from './free-text';
 export { default as MultipleChoice } from './multiple-choice';
 export { default as Poll } from './poll';
 export { default as PollAnswer } from './poll-answer';
 export { default as SubmitButton } from './submit-button';
+export { default as TextQuestion } from './text-question';
 
 export { renderBlocks } from './render-blocks';
